personal_projects/nn_stock/requirements.txt
python-dotenv>=0.5.1
pytorch
matplotlib
alpha_vantage
numpy